+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.15] - 2025-07-02
+
 ### Fixed
 
 - Missing callback for Admin Actions page output.
@@ -603,7 +605,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First Release!
 
-[unreleased]: https://github.com/26b/unbabble/compare/0.5.14...HEAD
+[unreleased]: https://github.com/26b/unbabble/compare/0.5.15...HEAD
+[0.5.15]: https://github.com/26b/unbabble/compare/0.5.14...0.5.15
 [0.5.14]: https://github.com/26b/unbabble/compare/0.5.13...0.5.14
 [0.5.13]: https://github.com/26b/unbabble/compare/0.5.12...0.5.13
 [0.5.12]: https://github.com/26b/unbabble/compare/0.5.11...0.5.12